+++ src/solvers/dgmulti/dg.jl
@@ -9,9 +9,11 @@
 mul_by!(A) = @inline (out, x)->matmul!(out, A, x)
 mul_by!(A::T) where {T<:SimpleKronecker} = @inline (out, x)->mul!(out, A, x)
 mul_by!(A::AbstractSparseMatrix) = @inline (out, x)->mul!(out, A, x)
+mul_by!(A::LinearAlgebra.AdjOrTrans{T, S}) where {T, S<:AbstractSparseMatrix} = @inline (out, x)->mul!(out, A, x)
 
 #  out <- out + α * A * x
 mul_by_accum!(A, α) = @inline (out, x)->matmul!(out, A, x, α, One())
+mul_by_accum!(A::AbstractSparseMatrix, α) = @inline (out, x)->mul!(out, A, x, α, One())
 
 # out <- out + A * x
 mul_by_accum!(A) = mul_by_accum!(A, One())
@@ -74,7 +76,7 @@ function create_cache(mesh::VertexMappedMesh, equations, dg::DGMultiWeakForm, Re
   @unpack wq, Vq, M, Drst = rd
 
   # ∫f(u) * dv/dx_i = ∑_j (Vq*Drst[i])'*diagm(wq)*(rstxyzJ[i,j].*f(Vq*u))
-  weak_differentiation_matrices = map(D -> -M \ ((Vq * D)' * diagm(wq)), Drst)
+  weak_differentiation_matrices = map(D -> -M \ ((Vq * D)' * Diagonal(wq)), Drst)
 
   nvars = nvariables(equations)
 
@@ -82,7 +84,7 @@ function create_cache(mesh::VertexMappedMesh, equations, dg::DGMultiWeakForm, Re
   u_values = allocate_nested_array(uEltype, nvars, size(md.xq), dg)
   u_face_values = allocate_nested_array(uEltype, nvars, size(md.xf), dg)
   flux_face_values = allocate_nested_array(uEltype, nvars, size(md.xf), dg)
-  if typeof(rd.approximationType) <: SBP
+  if typeof(rd.approximationType) <: Union{SBP, AbstractDerivativeOperator}
     lift_scalings = rd.wf ./ rd.wq[rd.Fmask] # lift scalings for diag-norm SBP operators
   else
     lift_scalings = nothing
@@ -108,7 +110,8 @@ function compute_coefficients!(u, initial_condition, t,
 
   # evaluate the initial condition at quadrature points
   @threaded for i in each_quad_node_global(mesh, dg, cache)
-    u_values[i] = initial_condition(getindex.(md.xyzq, i), t, equations)
+    u_values[i] = initial_condition(SVector(getindex.(md.xyzq, i)),
+                                    t, equations)
   end
 
   # multiplying by Pq computes the L2 projection
@@ -155,10 +158,10 @@ function prolong2interfaces!(cache, u, mesh::AbstractMeshData, equations,
   apply_to_each_field(mul_by!(rd.Vf), u_face_values, u)
 end
 
-function calc_volume_integral!(du, u, volume_integral::VolumeIntegralWeakForm,
-                               mesh::VertexMappedMesh,
+function calc_volume_integral!(du, u, mesh::VertexMappedMesh,
                                have_nonconservative_terms::Val{false}, equations,
-                               dg::DGMulti, cache)
+                               volume_integral::VolumeIntegralWeakForm, dg::DGMulti,
+                               cache)
 
   rd = dg.basis
   md = mesh.md
@@ -254,7 +257,7 @@ end
 
 # Specialize for nodal SBP discretizations. Uses that Vf*u = u[Fmask,:]
 function prolong2interfaces!(cache, u, mesh::AbstractMeshData, equations, surface_integral,
-                             dg::DGMulti{NDIMS, <:AbstractElemShape, <:SBP}) where {NDIMS}
+                             dg::DGMultiSBP)
   rd = dg.basis
   @unpack Fmask = rd
   @unpack u_face_values = cache
@@ -269,10 +272,10 @@ end
 # du[Fmask,:] .= u ./ rd.wq[rd.Fmask]
 function calc_surface_integral!(du, u, surface_integral::SurfaceIntegralWeakForm,
                                 mesh::VertexMappedMesh, equations,
-                                dg::DGMulti{NDIMS,<:AbstractElemShape, <:SBP}, cache) where {NDIMS}
+                                dg::DGMultiSBP, cache)
   rd = dg.basis
-  md = mesh.md
   @unpack flux_face_values, lift_scalings = cache
+
   @threaded for e in eachelement(mesh, dg, cache)
     for i in each_face_node(mesh, dg, cache)
       fid = rd.Fmask[i]
@@ -287,7 +290,6 @@ calc_boundary_flux!(cache, t, boundary_conditions::BoundaryConditionPeriodic,
 
 # "lispy tuple programming" instead of for loop for type stability
 function calc_boundary_flux!(cache, t, boundary_conditions, mesh, equations, dg::DGMulti)
-
   # peel off first boundary condition
   calc_single_boundary_flux!(cache, t, first(boundary_conditions), first(keys(boundary_conditions)),
                  mesh, equations, dg)
@@ -354,7 +356,7 @@ end
 calc_sources!(du, u, t, source_terms::Nothing,
               mesh, equations, dg::DGMulti, cache) = nothing
 calc_sources!(du, u, t, source_terms::Nothing,
-              mesh, equations, dg::DGMultiFluxDiff{<:SBP}, cache) = nothing
+              mesh, equations, dg::DGMultiFluxDiffSBP, cache) = nothing
 
 # uses quadrature + projection to compute source terms.
 function calc_sources!(du, u, t, source_terms,
@@ -371,7 +373,8 @@ function calc_sources!(du, u, t, source_terms,
     u_e = view(u_values, :, e) # u_values should already be computed from volume integral
 
     for i in each_quad_node(mesh, dg, cache)
-      source_values[i] = source_terms(u_e[i], getindex.(md.xyzq, i, e), t, equations)
+      source_values[i] = source_terms(u_e[i], SVector(getindex.(md.xyzq, i, e)),
+                                      t, equations)
     end
     apply_to_each_field(mul_by_accum!(Pq), view(du, :, e), source_values)
   end
@@ -381,29 +384,29 @@ function rhs!(du, u, t, mesh, equations,
               initial_condition, boundary_conditions::BC, source_terms::Source,
               dg::DGMulti, cache) where {BC, Source}
 
-  @trixi_timeit timer() "Reset du/dt" fill!(du, zero(eltype(du)))
+  @trixi_timeit timer() "reset ∂u/∂t" fill!(du, zero(eltype(du)))
 
-  @trixi_timeit timer() "calc_volume_integral!" calc_volume_integral!(
-    du, u, dg.volume_integral,
-    mesh, have_nonconservative_terms(equations), equations, dg, cache)
+  @trixi_timeit timer() "volume integral" calc_volume_integral!(
+    du, u, mesh, have_nonconservative_terms(equations), equations,
+    dg.volume_integral, dg, cache)
 
-  @trixi_timeit timer() "prolong2interfaces!" prolong2interfaces!(
+  @trixi_timeit timer() "prolong2interfaces" prolong2interfaces!(
     cache, u, mesh, equations, dg.surface_integral, dg)
 
-  @trixi_timeit timer() "calc_interface_flux!" calc_interface_flux!(
+  @trixi_timeit timer() "interface flux" calc_interface_flux!(
     cache, dg.surface_integral, mesh,
     have_nonconservative_terms(equations), equations, dg)
 
-  @trixi_timeit timer() "calc_boundary_flux!" calc_boundary_flux!(
+  @trixi_timeit timer() "boundary flux" calc_boundary_flux!(
     cache, t, boundary_conditions, mesh, equations, dg)
 
-  @trixi_timeit timer() "calc_surface_integral!" calc_surface_integral!(
+  @trixi_timeit timer() "surface integral" calc_surface_integral!(
     du, u, dg.surface_integral, mesh, equations, dg, cache)
 
-  @trixi_timeit timer() "invert_jacobian" invert_jacobian!(
+  @trixi_timeit timer() "Jacobian" invert_jacobian!(
     du, mesh, equations, dg, cache)
 
-  @trixi_timeit timer() "calc_sources!" calc_sources!(
+  @trixi_timeit timer() "source terms" calc_sources!(
     du, u, t, source_terms, mesh, equations, dg, cache)
 
   return nothing