+++ test/test_structured_2d.jl
@@ -95,9 +95,9 @@ isdir(outdir) && rm(outdir, recursive=true)
 
   @trixi_testset "elixir_advection_restart.jl with waving flag mesh" begin
     @test_trixi_include(joinpath(EXAMPLES_DIR, "elixir_advection_restart.jl"),
-      l2   = [0.00016265538283059892],
-      linf = [0.0015194298895079283],
-      rtol = 5e-7, # Higher tolerance to make tests pass in CI
+      l2   = [0.00016265538265929818],
+      linf = [0.0015194252169410394],
+      rtol = 5e-6, # Higher tolerance to make tests pass in CI
       elixir_file="elixir_advection_waving_flag.jl",
       restart_file="restart_000021.h5")
   end
@@ -210,14 +210,16 @@ isdir(outdir) && rm(outdir, recursive=true)
     @test_trixi_include(joinpath(EXAMPLES_DIR, "elixir_hypdiff_nonperiodic.jl"),
       l2   = [0.8799744480157664, 0.8535008397034816, 0.7851383019164209],
       linf = [1.0771947577311836, 1.9143913544309838, 2.149549109115789],
-      tspan = (0.0, 0.1))
+      tspan = (0.0, 0.1),
+      coverage_override = (polydeg=3,)) # Prevent long compile time in CI
   end
 
   @trixi_testset "elixir_hypdiff_harmonic_nonperiodic.jl" begin
     @test_trixi_include(joinpath(EXAMPLES_DIR, "elixir_hypdiff_harmonic_nonperiodic.jl"),
       l2   = [0.19357947606509474, 0.47041398037626814, 0.4704139803762686],
       linf = [0.35026352556630114, 0.8344372248051408, 0.8344372248051408],
-      tspan = (0.0, 0.1))
+      tspan = (0.0, 0.1),
+      coverage_override = (polydeg=3,)) # Prevent long compile time in CI
   end
 
   @trixi_testset "elixir_mhd_ec.jl" begin